ravif: Add `#[deprecated]` notice to `ColorSpace`
Commit e03108a ("Rename color space → model") (for #88) added/kept
hidden variants of `ColorSpace` and `with_internal_color_space()`
for backwards compatibility, to not need a semver-breaking release.
Unfortunately these were not annotated with `#[deprecated]`, making
users unaware of this change until they would (if ever) upgrade to the
next breaking release of `ravif` where such helpers are supposedly going
to be removed again.

Tell users about this with a `#[deprecated]` notice to guide them to the
new item names.