README: Add missing `https:` prefix in many `lib.rs` links
By missing `https:`, the [current crates.io page for `cavif`] and
[`ravif`] seem to link straight to files in the GitHub repository, which
doesn't exist.  lib.rs in contrast seems to link this correctly to its
own site.

For `cavif`, also mention that it goes through the `ravif` crate which
disambiguates the first paragraphs of both crates and makes it easier to
understand the total dependency hierarchy.

[current crates.io page for `cavif`]: https://crates.io/crates/cavif/1.5.6
[`ravif`]: https://crates.io/crates/ravif/0.11.11